soober roar!!!!!!!
this commit is basically just for adding the soober ability, the roar. this replaces the neutral monster's ability from classic where it runs towards the player. in the test level it might not be much, but imagine how challenging it could be in a tight section in a level... ouch.

currently the roar takes away 3 heart auras per hit (as opposed to roguts dealing only 1 per hit). i might change this in the future but for now it seems alright.

to-do for next time since ive decided not to do it tonight:
- add bg transition from title to main menu.
- finally get to updating the options button

- after these are done i should probably get to making proper sprites for the roar and roaring soober in the commit after lol. we'll see though