world 2 officially begins
ok so world 2 will not be lantern star anymore, but will instead be a fall themed world called sunshine falls. it'll basically be like a spiced up world 1 with floraisakus and perhaps even a new type of rogut??? speaking of roguts i reverted the spawning overhaul because it doesn't work in the levels for some (no) reason (but it works in the test level just fine???) so until i figure something out we'll deal with the spawning mechanics another time.

back to world 2 talk, the graphics and shadow stuff (tileset updated!) are still very much a work in progress, but i guess this'll do for now.