removed unnecessary folders that won't matter for the demo (aka im lazy and don't wanna work on lantern star for the 2025 demo)
in other news FNF UPDATE DROPPED IM WINNING TONIGHT. in other other news i started using trello again so i more or less have an idea of what i need to work on before the demo drops. my prediction for the drop for all the peeps in the future reading this to be like, "wow this guy was wrong"? demo coming out january 25th 2025 to celebrate 6th gooter anni